I: remove babel with typescript
I don't like typescript, but I plan to use this lib with Angular and
therefor I find it more convenient for the developers to have type hints.
Not sure if that was a bad idea. It could cause a lot more work, with
only a little bit of advantage. But, let's try it.